miRkwood: Add candidate merging condition
Candidates can overlap each other or be included into one another.
We want to process differently both cases.

In the case of overlapping, we take the one with the best MFEI,
like before.
In the case of included, we want to take the bigger one if its
MFEI is lower than -0.8, otherwise the best MFEI.

- Split predicate is_to_merge in two methods:
  is_overlapping and is_included
- Add assertions to these methods: die if positions are unordered
  or if positions are not initialized
- Implement in merge_candidates the aforementionned behaviour
- Add unit tests for the new predicates.