miRkwood: Text label triggers checkboxes in inteface.pl
I want the checkboxes in the interface page to be selected
when I click on the label text, in order to more easily select.

This is achievable using the <label> element with the for attribute
referencing the input ID. This is also important for accessibility
purposes.

See <https://developer.mozilla.org/docs/Web/HTML/Element/Label>
for more informaiton on the label tag.