miRkwood: Text label triggers radio buttons in resultsWithID.pl
I want the radio button on the results page to be selected
when I click on the label text, in order to more easily select.

This is achievable using the <label> element with the for attribute
referencing the input ID. This is also important for accessibility
purposes.

See <https://developer.mozilla.org/docs/Web/HTML/Element/Label>
for more informaiton on the label tag.

Add IDs to button as well.