miRkwood: Fix bug in candidate merging if not candidates
miRwood was failing with the following error message:
"Can't use an undefined value as a HASH reference
at lib//PipelineMiRNA/MainPipeline.pm line 321"

The offensive 321 line is:
my %reference_candidate = %{ $candidates_array[0] };

Where candidates_array is the given array of candidates.
If said array is empty, there is no element to dereference,
hence the failing.

We solve this bug upstream, by testing if the candidates array is empty.
If not, merge_candidates is called ; if so, we just use an empty hash.

(We really need to find a way to output die messages in the log file).