miRkwood: Redirect STDERR to log file
By default, Perl errors are stored in the Apache log file,
which is not convenient at all when something goes wrong and we
are painfully scouring the log.log file.

Using carpout, we redirect the STDERR to the job log file.
(We considered creating a mirkwood log file, but decided against for now)

See <http://perldoc.perl.org/CGI/Carp.html#REDIRECTING-ERROR-MESSAGES>