miRkwood: Add centralized logging verbosity handling
We use Log::Message::Simple to log stuff.
Debug messages make it to the log file if and only if
the method call is flagged true.

We used to hardcode the flag in each call, which was particularly evil
as there was no easy way to turn it on or off.

Add a global variable (accessed through a property-like method)
at package level

Access this variable for every debug call.