miRkwood: Select only sequences with MFEI < -0.6
As a user, I want to avoid processing candidates if their
MFEI is above -0.6, because it is assumed they are not interesting,
in order to save computation time.

We implement this selection during the main pipeline,
just after retrieving the output of RNAstemloop,
just before merging the candidate.

On our example data, with both Randfold and alignments,
this cuts down the processing time from 50s to 38s.