miRkwood: Remove unit test for get error page
get_error_page now relies on get_HTML_page_for_content,
which relies on finding the static files for the menu and header.
In the unit test environment, there is no guarantee to find those,
as they rely on the local configuration (pipeline.cfg)

Since there is no chance this will get sorted this out soon, and
since the tests brings next to no business value, remove it.