miRkwood: Fix bug in ASCII hairpin with mature on other side (redux)
The method make_hairpin_with_mature builds positions the mature on the hairpin
The mature can be located on either 'side' of the hairpin.

To determine if the mature was to be located on the 'other' side
(meaning: the opposite), the left boundary was compared to the
length of the top row of the hairpin, and it works on many cases
where the mature is quite clearly on the other side.

But the right boundary could still overflow on the other side.
What makes sense here is to compare to the right boundary.

Add unit test covering this case.