miRkwood: Divide string subtitutions in two operations
The method substr is used twice in the same line, first in selection mode,
and then in replacement mode (with the result of the former).
This makes the method harder to understand, and also possibly harder to
debug: if something goes wrong on that line due to substr, there is no
way to know which one is the culprit.

Use temporary variable to store the result of the first substr
(with a meaningful name), and use it in the second.