miRkwood: Remove non-Fasta format bug
miRkwood is suppose to return en error if a non FASTA file is provided.

The method to check this was way too permissive, and allowed
non ATGCU characters.
(This was detected as our Shuffler does not like such sequences either).

Edit the is_FASTA method to pass only on ATGCU.
Add unit tests.