miRkwood: Adjust exonerate positions
The position information given by exoerate is not 'positions'
per se, but rather 'boundaries'.
When the alignment of the sequence is between boudaries $x$ et $y$,
it means positions $x+1$ and $y$.
The same goes for the database alignment.

Edit the parsing method to add this offset.
The special case is for the database ('query') sequence when
it is reversed - which we will eliminate at some point anyway,
but it least it is not confusing in the mean time.

Amend unit test accordingly.