miRkwood: Remove cruft from CGI script results.pl
There was a lot of cruft left here, leftovers of old processing.
For example, we were writing the sequence on disk, and immediately
after reading it back into memory (this was useful at some point as
we were running some Bash scripts, but not anymore).

Remove the reliance on a SED command, rather using a pure Perl method.