miRkwood: Add custom-built exonerate binaries
miRkwood relies on the exonerate software as a 3rd party dependency.
<http://www.ebi.ac.uk/~guy/exonerate/>
Exonerate can easily be fetched through the APT repositories.

We identified that exonerate, when used in DNA mode, processes
the reverse complement the query sequences (in our case, miRbase).
This is no good for us, because it makes no sense in our use case
and uses double the time.

We thus use a custom built exonerate which never processes the reverse
complement of the query sequence.
The change is made in the file /src/database/fastapipe.c, line 40:

-    fasta_pipe->revcomp_query
-        = (query_fdb->alphabet->type == Alphabet_Type_DNA);
+    fasta_pipe->revcomp_query = 0;
+//        = (query_fdb->alphabet->type == Alphabet_Type_DNA);

The built binaries, for both i386 and x86_64 architectures,
are checked in the version control as a third-party dependency.