miRkwood: Make hairpin with mature in ASCII mode
As a user, I want to position the mature on the hairpin,
even not in HTML format.

Add another parameter to method make_hairpin_with_mature,
to chose the way the mature is positioned and defaults to ascii.

The HTML does what is currently done: surrounding the mature
with an HTML <span> with CSS.

The new ASCII mode puts the mature in uppercase.

Add unittests for the new behaviour.

Refactor unit tests to use data files.