miRkwood: Fix bug in is_fasta predicate
The method is_fasta is meant to detect whether a file matches
our requirements for FASTA files as miRkwood input.

The check was failing on the "humongous Hélène file", though nothing
seemed wrong with it.
As it was a "global" method, checking using a regex the entire file,
there was no easy way to determine where the method was complaining.

Here we rewrite the method to operate on a line per line basis.
We add two companion methods, one checking conformity to a FASTA
header and the other to a FASTA line as accepted by miRkwood.