miRkwood: Fix bug in candidate sorting before merging
To merge the candidates, they have to be sorted.
Though this was done when both strands were used, it was not done
when used on only one strand. This was not detected until now
because in most of the time candidates are already ordered.

(This was caught by the preconditions in is_overlapping)

Add sorting to candidates array before merging.