miRkwood: Remove chdir during miRdup execution
For some reason, we need to chdir to the mirdup model directory
before executing the command.

We were doing this with a Perl chdir, which is particularly evil.
This can break assumptions on paths built with constructs like __FILE__
and there is no good reason for that anyway.

THIS was provoking an evil bug which prevented from retrieving the
configuration file after the chdir, as its path was defined at startup,
resulting in failure.

Change to cd in the shell and not in the Perl.