miRkwood: Revamp configuration handling
miRkwood relies on a fair number of paths which are stored in YAML
configuration files.

These files were accessed every time a value was needed, which was
preventing any more sophisticated way of dealing with configuration
(like injecting it at runtime) and maybe not very inefficient either.

Here we revamp the all thing to store the configuration in the
PipelineMiRNA root module: a global variable for the path to the
config file and another for the actual config.
Methods allow to get and set them.
They are initialized at start time, the FILE with the default
value and the config with the deserialized config file.

Afterwards, programs access config values in memory.

This was initially done to circumvent an evil bug where
the path to the config file was not accessible anymore during
the application run (because of a directory change), which was
solved at the source in the mean time.
Nevertheless, the application is more robust that way.

Programs configuration is still flaky though. We need to continue
its revamp to target all needed binaries directly.

Genreal Configuration could do with separating all that is webapp
and all that is pipeline.