miRkwood: Remove Energy computation as a posteriori test
The Energy computation used to be one of the optional aposteriori tests.
But in fact, we need the MFEI in all cases way before, as a criterion
for merging the candidates and optionally filtering the crappy ones.

The all Energy thing was thus computed again for no reason. Removing that.