miRkwood: Do not display the alignment section if not selected.
The user may chose to not perform the alignments.
In that case, we do not want the "Conserved miRNA sequences"
section to be displayed, as the user could be misled into
believing none were found (when we did not even try).

We achieve this by using the configuration setting in the run_options file.