miRkwood: Update table icons
The miRkwood results table makes use of icons
for some elements: a star, a check, and a cross.

We want to have a good-looking, unclutterred set of icons,
with a coherent look and feel.

We decided to use icons from the NounProject
<http://thenounproject.com/term/check-mark/16645/>
<http://thenounproject.com/term/close/19521/>
<http://thenounproject.com/term/star/3307/>

They were altered on Inkscape to use the Tango! palette
and converted to PNG.

Edit Javascript to use the new icons.
Add real alternative texts.