miRkwood: Fix bug in position extraction from RNAstemloop output
Various information are extracted from RNAstemloop output:
the name, the positions, the sequence and the dot-bracket.

Name and poistion are isolated based on the two underscores.
But the name of the sequence might contain two underscores,
screwing up the parsing completely.

For example, with a sequence named "random_seq_from_cds__no_71457"
the parsed name was "random_seq_from_cds", and the position 'no_71457',
ignoring the positons.

Here we use a regular expression which copes with with such possibilities
in the sqeueqnce name.