miRkwood: Move configuration handling and store job name in it.
The job title used to be passed along to webpages in GET, but was not stored
anywhere and thus lost.
We want to have it displayed in any case.

Here we add it to the job configuration file.

To perform this, we refactor the job config handling,
so that the configuration is stored earlier. This is useful because this way
the main_entry does not need to take many parameters, since it may retrieve
the information form the job config file (something it had to do later on anyway).