miRkwood: Change method to isolate sequence name for a fixed length
We need to retrieve from the FASTA header a sequence name.

The former way to do this was to split on spaces and retrieve the first element.
This had unexpected side effects, like sequence starting with a space,
or more often isolating a common name like 'random', resulting in duplicate
names.

Here we change to shorten at 30 characters.

Update unit tests.