miRkwood: Revamp positions handling (and fix evil bug)
When sequences are processed with the opposite strand,
positions are inverted. We restore the 'good' position right away,
but it turned out that the 'wrong' one was the one which serialized,
and thus used on all the displays.

This was busted because it seemed as if candidates were not merged properly ;
but in fact they were, it were the positions displayed which were wrong.
(which, somehow, is way more evil).

Here we change the ay these positions are serialized.
We used to get the merged position from the RNAlfold file,
and the start/end were found by splitting.

Here we repurpose the 'strand' file (holding the strand information)
to hold also start and end positions (the 'good' ones).
At parsing time, the start/end are retrieved.
They are concatenated too into a global position,
mainly for backwards compatibility (we should actually make it a class
method , but well.)

Fix unit tests as well