miRkwood: Revamp short name handling
A long time ago, we had to sanitize sequences names
because they were semantically used, and as such might screw up
with hash keys or filesystem entries.

A while ago we removed any hard dependency on this name, relying
on autogenerated identifiers.

Now the huge names are only a problem when displayed in the results
page.

Here we change that behaviour to parse and store the full FASTA name.
It is only when converting to pseudo XML and to GFF that
we use the shortened name, as well as when needing it for files downloaded.