miRkwood: move finish file writing to pipeline execution
To determine whether the job has finished,
we write at the end a file 'finished' on disk.

We used to do that in the CGI script, when returning from the spawned Perl process.
This turns out to be very fragile -- randomly the file does not get written.
Debug did not help much.

We move it at the end of the Perl method, which should work.

We might want to invert the logic here (testing the presence of a file
which is unlinked at the end), in order not to leave any useless traces.