miRkwood: Add option to not generate VARNA images
Generating images using Varna takes some time.
In some cases we cannot afford it.

In a previous commit we laid the ground for this, and finish it here.

Add setting it in configuration file creation
Add option in bin script mirkwood.pl
Update CGI script to default to true
Prevent serializing a non-adress

Note that the reports (HTML and ODF) do not take it into account.