miRkwood: fix bug in MFEI computation with a no GC sequence
The MFEI is computed by dividing the MFE by the GC content.

This crashes in the (albeit unlikely) case where the GC content is null,
ie with a sequence with no G nor C.

In that case we decide to set the MFEI to 0.

Add unit tests catching that case.