miRkwood: Rename position_start in start_position
start_position is the correct term to use in this context.
The name of the field is directly used in output in the CSV export.
It is easier and more consistent to change it everywhere.

Rename all instances of position_start by start_position.
Idem for position_end --> end_position

Amend unittests and test files accordingly.