miRkwood: Move configuration files to package level
It has always been a bit awkward to have the config files
right with the Perl modules.

A while ago we moved the code handling them to the PipelineMiRNA module.
From that point on, it had made no no sense to keep the files one
level above.

Move the cfg files up.