miRkwood: Heavily revamp configuration handling
Configuration handling has always been awful in miRkwood.
Duplicated paths, lots of things to change at deployment time,
variable names which did not make much sense...
And as usual, a too tightly coupled WebApp and CLI application,
which make the application hard to package without the web cruft.

Here is a (heavy) stab at rationalising it.

- Split configuraiton file in web-related and non-web-related
- Add to main module accessor to this new config
- Move to new module WebPaths all web-related configurations
- Replace js/css config values by one, html_root, and infer
  the former form the latter
- Infer most of the mirkwood config from mirkwood itself:
  data, scripts and lib all live in the same place, and there is no
  reason to be otherwise (hence no rationale from separating all three)
  Moreover, miRkwood lib knows where it is, and it can thus infer
  where are the others
- Rename unclearconfig variables (in the web config)
- Rename a method or two while we were at it
- Update all CGI scripts accordingly

Now we have a much better organized codebase.
Almost nothing relies on WebPaths module apart from WebTemplate and
the CGI scripts - which is as it should be.
Only left are related to the image handling:
- OpenDocument, but the behaviour there should be moved to
  the CGI script calling it ;

- Candidate XML export. This underlines the fact that the XML export
  only makes sense in the WebApp context. This should be spinned to
  a different module which could be cut at packaging time.
  As we were thinking of objectify all of this, and making a whole
  hierarchy of Results Formatters/Exporters (CSV, GFF, etc.),
  this could play nicely together.