miRkwood: Fix functional tests
We disabled the generation of one of the CT files,
wo we have to delete it from the expected results.

For some reason the serialization order in the YML file
is not the same as it was before. Change that too.

Add forgotten symlink which would have made the tests fail elsewhere.