miRkwood: Fix evil bug in is_included and is_overlapping
In order to safeguard the methods is_included and is_overlapping,
we added a check to ensure each one of the four parameters were
appropriately unloaded, and dying otherwise.

This resulted in a very evil bug, where the method died in case one
of the parameter value was Zero - as it was evaluated as false.

I cannot find words to desribe the horror this inspires me
(and it might not be the place for that anyway).