miRkwood: Add module to deal with Clusters
We want to use miRkwood in a sRNAseq context,
with an additional step analysing a BAM file
to retrieve peaks and extracting candidate sequences
which will be used as input for the rest of the pipeline.

Add module Clusters.pm (with acompanying test file)
which will take care of that.