miRkwood: Add method get_faidx_file
To identify the islands we need an indexed genome file,
which might not have computed.

Add method get_faidx_file which looks for the index file
based on the classic naming pattern (x.fai) and indexes
the genome file if needed using the samtools.

Add unit tests covering both cases.