miRkwood: Delete temporary dummy test files
We have no reason to keep these files around, better delete them,
and it is better for tests to cleanup after themselves.

They were also causing an issue with the CI server.
Because of some quirks in our setup, tests are executed twice,
once for test results, the other for code coverage.
Having the files around for the second round defeated some tests
which messed up the code coverage.