miRkwood: Add method get_clustered_sequences_from_bam
As a user, I want miRkwood to take a BAM (with a genome),
find loci in it and run the rest of the pipeline on it.

Here we (almost) finish the implementation of this.
The method get_clustered_sequences_from_bam is to be the
main interface for the Clustering part.

We still need to extend the clusters though!