miRkwood: Fix functional tests after rename
The functional tests include a test of the BLAST filtering,
for which a database must be given.

The Arabidopsis thaliana database was renamed at some point,
but the functional tests were never fixed − they kept working
locally though, and the CI server did not execute them for a while.
(Now it does in a dedicated job).

Update test to use the new database name.