miRkwood: Add method to execute tRNAscanSE
We want to use tRNAscanSE to mask noisy RNAs from our input files.

This method runs tRNAscanSE, which is a bit tricky.
PATHs *must* be set in order to use it (otherwise it does not find
some Perl underlying library).

Solution here is to store in the config file the path to tRNAscanSE
and to set it before running the command.