miRkwood: Revamp BLAST filtering
The Blast feature was totally not doing the right thing:
it filtered the sequences with a hit, instead of masking
said hit in the sequence.

This fixes this major feature fail, by introducing
the mask_sequences method to mask a given sequence with positions.

This paves the way for the tRNAscan-SE masking (but not tonight).