miRkwood: Remove InstallGuide from Dist::Zilla config
This Dist::Zilla plugin is supposed to generate the INSTALL file.
It used to work but on the latest version it fails because it
cannot find Build.pl nor Makefile.pl (though Dist::Zilla is supposed
to generate these files...)

As it was not terribly useful anyway, let's just remove it.