miRkwood: Dying in web context
In the web context, there are a lot of cases where we have to die.
This results in either a Software Error (if FatalsToBrowser are enabled)
or in "Internal Server Error".

Both cases are super frightening.

Here we introduce a simple die method which wraps around the printing
of the error page and the dying, and replace all 'die' instances with it.

This results in much nicer error messages.