miRkwood: Graceful handling of ODF absence.
The ODF::lpOD library is necessary when importing the OpenDocument module.
But we want to be able to use miRkwood without it, since in most cases
it's not a problem.

The problem is that 'use' works at compile time − thus no export
could be used withou importing this library.

Using 'require' solves the issue, since it works at runtime.
We also use the we dying method to better explain the issue to our poor users.