miRkwood: Use temporary directory for RNAmmer execution
RNAmmer needs to write stuff on disk, and thus needs permission to do so.
By default, it chose the current directory. Not sure where that is in the
context of the web server, but clearly the user www-data does not have
permission to write wherever that is.

Fortunately RNAmmer has an option to use a given folder as the temp folder.

Use module File::Temp which allows to have a temporary directory.
Use temporary directory with option -T in RNAmmer command line.