miRkwood: Refactor masking code
We want to decouple as much as possible the modules other than
MainPipeline from the job paradigm.

It was not ideal to have the job directory given to the masking methods,
as they only care about the sequences there.

Plus now we have a dedicate folder to hold the text files, which may
avoid future clutter.