miRkwood: Revamp internal program configuration management
Programs paths were managed using as much global module variables,
set from a dictionary set from a config file.

This was a nightmare, as it was cumbersome to add more programs,
while it is easy to just use a dictionary set from the default config.

Doing precisely that.

This should allow uw to more easily override defaults.