miRkwood: allow for variable local path in programs configuration
For several reasons (have to be distributed with miRkwood, super annoying to install),
some programs need to be found with miRkwood itself
(while the others are expected in the PATH)

We used to have hardcoded paths in the configuration file, which was super evil,
especially considering that miRkwood should be able to locate these programs alone
since it just KNOWS where it lives.

We solve that by allowing a special syntax in the configuration file
(vaguely reminiscent of a templating system − or not).
This special syntax is replaced at runtime by miRkwood location.

Note that configuration overload might still be useful ;
but still that just makes so much sense.