miRkwood: Extract method get_Vienna_program_path to fix bug
miRkwood in web mode was failing with
fileparse(): need a valid pathname
at /home/data/bio1/www/cgi-bin/mirkwood/web_scripts/../lib/PipelineMiRNA/Programs.pm line 30

That was because the logic inferring the Vienna program location based on
the RNAfold executable specified in the configuration file was only working
when referring to execs in the PATH, not by their full path.

This patch allows for both logics, and is extracted to its own method
for clarity.