miRkwood: Fix bug with alignment in OpenDocument.
In the ODT file, alignments were not working anymore:
the middle line was not properly indented:

query            19 uacucccucugucccauuacag 40
|||||||||||| ||| || ||
miRBase           1 uacucccucuguucca-uaaag 21

This bug was because of a careless change from '' to qw{},
jsut because Perl::Critic happens to recommend it.

Note that this code is duplicated in OpenDocument and Candidate module.