miRkwood: Safeguard candidate info deserialization
Make candidate creation through information deserialization
fail if the given path does not point to a file.

The Results::deserialize_results method loops in the intended
directory and deserialize everything it finds.
This could be very problematic, as Candidate objects are created
and fed even with an empty data hash.
This was highlighted on the CI server, where a .svn directory is present.

As a first stop-gap measure, the creation dies if the
serialiazed path is not a file. This should solve the
current symptom but is not enough.