miRkwood: Safeguard candidate info deserialization
Make candidate creation through information deserialization
fail if the deserialized object is empty.

Follow-up to the previous commit.

The Results::deserialize_results method loops in the intended
directory and deserialize everything it finds.
There could be anyting there, and we do not want to create objects
empty-initialized based on spurious files.