miRkwood: Fix functional tests expected data
In the functional tests we compare the filetree
of expected values (versionned) and output.

This was failing because the serialization order of Candidate
in the YAML files was different (probably linked to the humongous
refactoring, but who knows).

Change expected data to match the new output. But these functional
tests are shaky anyway, we should think more on that.