miRkwood: Refactor Pipeline into several objects
Refactor the huge MainPipeline module into two smaller OO modules.

MainPipeline.pm was this huge, unreadable and unmaintainable
piece of 850 lines of code loosely related, with a bunch of methods
carrying the same parameters around forever.

In this commit we refactor this completely.

Pipeline is a OO module around the concept of a miRkwood pipeline job
(and this around the job directory)
The BamPipeline module inherits from it and only redefines the
method to initialize the sequences, as it shares most of its code.

SequenceJob is a 00 module around the part of the miRkwood pipeline
runnning on a single sequence.

There is still much work to be done, but this is a start.