miRkwood: Have jobs return their results to caller
This architecture of writing results on disk was an
artifact of when mirkwood was a collection of Perl scripts.

Now we should have subjobs return their results to the caller:
- CandidateJob will return a candidate object to the SequenceJob
- SequenceJob will aggregate them and return the collection
  to the Pipeline

This will allow us to do more powerful things, but firstly
have the serialization process saner.

Ultimately we will completely hide the folder tree system
behind good abstractions. Yay!