miRkwood: Revamp candidate serialization
In the previous commit, the results of the jobs were returned
to the Pipeline module, as a (would-be) Candidate object.

With that we can now revamp the candidate serialization process:
- the candidate object is created in the CandidateJob
- the Pipeline retrieves the candidates by return call,
  and ask them to serialize themselves in the right directory

The CandidateHandler module still acts as a middleman here,
though with reduced responsabilities. Ultimately this middleman
should go away.

Further refactoring is needed to get rid of some useless
writing/parsing of information at the candidate level.

Update CandidateHandler unittest accordingly.

(Also, alter functional test expected file, for some reason
This is really super fragile).