miRkwood: Major revamp of CandidateJob
CandidateJob was annoying because it was writing all it knew on disk,
to parse it again and return it (leftover from the script architecture).

Here we no longer write on disk anything unless we need it for
the a posteriori tests.

This implied a humongous refactoring only to make tests still pass.