miRkwood: Fix functional tests
The functional tests were failing when binary files differed,
in this case the files generated by VARNA - the structures were
the same but the images were oriented differently.
It is unclear why this is so - VARNA should be in the same version
than before.

In any case, binary files difference is fragile, and we are mostly
interested in testing the presence of these files.

Grep-ing out the 'Binary files X and Y differ' lines.

Remove the -q flag of diff, as it hides the distinction between
binary and non-binary, and because we do not care if diff
spits out a big diff: the wc will be different from 0 in both cases.