miRkwood: Add a new pipeline in which the user provides alignment data in the form of a BED file.
BAMpipeline.pl is the start page for this new pipeline.
The user must provide a BED file and choose a model organism
(currently only A.thaliana is available) or enter a reference
sequence.
Currently the form works well but the following is not yet
implemented.
Genome (fasta format) and annotations files (tRNA, rRNA, mRNA...
and also known miRNAS) in GFF format for A.thaliana are stored
in mirkwood/cgi-bin/data, in directories genomes/, annotations/ 
and miRBase/.