miRkwood: Add a function to turn the relative positions into absolute positions for a Candidate.
This function is called only for the BAM CLI pipeline. It's 
not easy to store directly absolute positions since the 
function which does that is also called for the genomic
pipeline and in this case the positions are already absolute.

So in the BAM CLI pipeline positions are turned into
absolute positions after all treatments and before printing
the reads clouds and creating the YAML files.