miRkwood: First attempt of a new score calculation for new miRNAs in the Web BAM pipeline.
Currently the stability of duplex has not been calculated, 
as we don't have the position of the duplex.
Also see the part of component 'alignment' in the calculation:
in the genomic pipeline there is one star if there is
alignments, and another star if validated by miRdup.
Should we only keep the mirdup part ?