miRkwood: Now known miRNAs are treated in run_pipeline instead of compute_candidates.
When the known miRNAs were treated in compute_candidates, 
the BEDPipeline object was larger because of the sequences
that are stored in $self->{'sequences'}, so when there is
a lot of known candidates it could lead to a memory crash.

Undef $self->{'basic_known_candidates'} after known
miRNAs are treated. Even if this variable is not big, 
it is not used anymore after that point.