miRkwood: change the threshold for the minimum nb of reads to keep a locus.
Previously this number was always set to 10.
Now we calculate the average read coverage (which is the
genome length divided by the number of reads). Then the new
threshold is calculated by dividing the locus length minus 20
by the read coverage. If this number is below 10 we keep 10
as the threshold.

This is to prevent having too much loci when the read 
coverage is important. In such cases there is a lot of
noise so we have to be more selective.