miRkwood: Now miRkwood is REALLY and completely launched chromosome by chromosome.
In the previous version, the processes to detect trains, 
spikes and loci were done in 3 successive loops on the
whole genome. 
Then the process to detect candidates was done chromosome
by chromosome too, but the hash with the loci data contained
loci data for ALL chromosomes, which was useless and needed
a lot of memory.

Now, the pipeline is really run chromosome by chromosome, so
for each chromosome the trains, spikes and loci are
calculated, then we look for hairpin candidates, then the
posteriori tests are made. Then the results are stored, the
variables are erased and we consider the following chromosome.
It needs much less memory.

However for now we can easily come back to the previous version
if needed, by commenting/uncommenting out a few lines in
BEDPipeline.pm (in function run_pipeline) and in 
HairpinBuilder.pm and ClusterJobSebastien.pm (look for
'#VOIE2').