miRkwood: Now the optional fields (for headers) only contains 'shuffles' and 'alignment' (that is, the only really optional fields).
The variable headers (such as reads, precursor_name...)
are determined in a more appropriate place (see modules
HTMLExporter, CSVExporter and daughters).

mfe, mfei, amfe and reads are always added in the list of
(potential) headers in get_basic_informations : if there is
some data in these categories, they will be added in the
'basic' informations. Elsewhere nothing particular happens.