miRkwood: Store in a new txt file (bed_sizes.txt) the number of reads and unique reads in each BED file.
These numbers are displayed on the main results page. Now
they are simply read in the txt file instead of being
calculated when the page is loading. Thus the page is
displayed more quickly.

Moreover once the bed_sizes.txt file is created we don't
need to read the BED files again so we can compress them
and save space (not done yet).