miRkwood: Fix a bug about displaying the Candidate card.
The identifier has to be among the data used for the display,
even if it is not displayed itself. It is needed to retrive
the candidate data and make the page.

So, as it doesn't seem easy to display the identifier
(which for known miRNAs contains the miRBase name) onlu
for known candidates and not new candidates, the miRBase
name is also stored in another attribute (precursor_name)
which is set only for known candidates.