miRkwood: Light change in the calcul of read coverage.
Reads coverage is taken as the genome size divided by the
number of reads, and for the genome size we only looked
at the last read position.
It's a fair approximation if the first read position is
around 0 or at least small enough compared to the last
read position, but it's not really accurate if the first
read starts at a quite high position.