miRkwood: Add a safety so that miRkwood won't crash if the CT file is not created.
- abinitio pipeline : the CT is used to mask the sequence
with N (create the seqWithN.txt file). If the CT is not
available, just copy the sequence (with T instead of U)
in this file.
- smallRNAseq pipeline : the CT is used to add the
'[[[..]]]' on the read clouds (to indicate the positions
of the miRNA and the paired sequence) and to calculate
one part of the read distribution score (whether there is
a star or not). If the CT is not available, just don't
show the line with '[[[..]]]' on the read cloud, and the
criterion about the existence of a star will always be 
set to 0.