miRkwood: Change the way we filter out reads according to the given gff.
Previously we filtered out CDS and t/r/snoRNA given in gff
files that had to be stored in a particular folder.
Now we can give as many gff files as we want (with the 
full path) and the BED file will be filtered according
to all these gff files.

Some changes had to be done in bed_sizes.txt and summary.txt.

Currently these changes only concern the CLI version (which
means that the Web version doesn't work currently).