miRkwood: Delete option db in the config file.
It was supposed to be used for Web version when the
user gives a reference sequence. The options 'db'
would have allowed the user to choose a reference
organism for the filters (like it is made with the
abinitio pipeline) but indeed we won't allow the
user to enter their own reference species. For that
purpose the user must use the CLI version.