miRkwood: Update miRBase fasta file used for alignments.
Some months ago I merged Mohcen's and Sylvain's scripts
to create an only script that can create this fasta file,
but I forgot to update the file...

The previous file has 'T' and no 'U', but it seems that
it's not necessary. Anyways the alignments are displayed
with 'U' so with 'U' in the miRBase file we don't have
to replace the T's.