miRkwood: Version a file containing the list of species from Viriplantae and the number of miRNAs precursors that miRBase known about them.
This file is obtained with a simple copy-paste from miRBase
( http://www.mirbase.org/cgi-bin/browse.pl ) (expand all
and then copy all what is between Viridiplantae and Viruses).

We need this file to create the MirbaseFile.txt file that
contains the miRNAs from miRBase and that is used by
miRkwood (more precisely, RNAcomp/piccolo), in fact only
the first column with the species names is necessary so
don't bother with the number of precursors.