miRkwood: Add a js file for some functions that are taken from /bio1/www/html/scripts/bioinfo.js (which is a common file for all softwares).
If we use this file the anchors links won't work.
Moreover miRkwood is not a dynamic website so most of
the functions in the bioinfo.js file are not useful 
for miRkwood.
I just copyed and pasted the functions useful for
the arborescence, for the header (change the color
of the active tab) and for the title.