miRkwood: CLI versions: change the way the results files are created.
Until now, the basic yaml files were used only for Web version,
and for CLI versions we used to read all files in YML folder.

Now we use the basic files to create the results, but for that
I needed to add some informations in the basic yaml files.
In fact we need to keep a lot of features for each candidate,
but we don't keep the bigger ones (reads, CT).
We also need the alternatives candidates, whose corresponding
feature in the hash is quite big, but in the next commit
I'll clean it to keep only what's needed.