miRkwood: clean the 'alternatives' features.
The 'alternatives' feature is needed only to create
the alternative files (function alternativeCandidatesAsVienna
from Candidate.pm) and only need the name, MFEI, sequence
and structure stemloop.
Now the hash with alternative structures has the name
concatenated with the MFEI as keys, and the only other
features that are kept are the sequence and the structure
stemloop.