miRkwood: Create a function which count in the BED file how many times the miRNAs are aligned.
This is calculated in the 'basic' informations and added
only in the basic_candidates.yml file, not in the YML
for each candidate.

That's why I needed to create the results from the
basic yml even for CLI version.

Add a column in the HTML table to show this number.